requirements.txt
Django
psycopg[binary]
djangorestframework
whitenoise
python-dotenv
django-allauth
anthropic
openai
replicate
fal-client
langchain
langchain-google-genai
langchain-openai
celery
redis
django-celery-results
django-redis
sqlalchemy
boto3
django-storages
sentry-sdk
PyJWT
cryptography
httpx[http2]
uvicorn[standard]
aiofiles
requests
paddle-python-sdk
google-auth
google-auth-httplib2
google-auth-oauthlib
gunicorn
pydantic
psutil
bleach
pillow
mutagen
pytest
pytest-django
tiktoken
python-gettext